refactor: remove parquet.rs, move TableKind to pg_parquet.rs
- Deleted src/sync/parquet.rs (~2000 lines of Arrow builder code)
- Moved TableKind enum to pg_parquet.rs
- Updated all references to use pg_parquet::TableKind

Amp-Thread-ID: https://ampcode.com/threads/T-019c0c14-5814-770b-9097-05650562e8f5
