+++ src/sync/mod.rs
@@ -1,7 +1,6 @@
 pub mod decoder;
 pub mod engine;
 pub mod fetcher;
-pub mod parquet;
 pub mod pg_parquet;
 pub mod replicator;
 pub mod writer;
